dictionary app with HTML, CSS, JavaScript
a website that provides comprehensive definitions for words in the English language. It offers a search engine to easily find definitions for words, as well as a variety of other services such as word of the day, audio pronunciations, and more. It is the go-to source for anyone looking to learn more about words and their meanings.